app/requirements.txt
# Core Django and web framework
Django==3.2.25
asgiref>=3.3.4,<4.0.0
sqlparse>=0.4.1,<0.5.0

# Authentication and security
msal>=1.10.0,<2.0.0
PyJWT>=2.0.1,<3.0.0
cryptography>=41.0.0,<48.0.0
cffi>=1.14.5

# HTTP and networking
requests>=2.31.0,<3.0.0
urllib3>=2.0.0,<3.0.0
certifi>=2023.0.0
idna>=3.0.0,<4.0.0
chardet>=5.0.0,<6.0.0

# Data processing
PyYAML>=6.0.1,<7.0.0
python-dateutil>=2.8.1,<3.0.0
pytz>=2023.3

# Utilities
six>=1.15.0,<2.0.0
pycparser>=2.20,<3.0.0

# GUI framework
PySide6>=6.5.0,<7.0.0

# Google API client libraries
google-api-python-client>=2.0.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth-oauthlib>=0.5.0,<1.0.0
